refactor: merge namespace-members into namespaces
