+++ src/app/app.module.ts
@@ -12,7 +12,6 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { UserRoleModule } from 'src/user-role/user-role.module';
 import { ResourcesModule } from 'src/resources/resources.module';
 import { NamespacesModule } from 'src/namespaces/namespaces.module';
-import { NamespaceMembersModule } from 'src/namespace-members/namespace-members.module';
 
 @Module({
   controllers: [AppController],
@@ -27,7 +26,6 @@ import { NamespaceMembersModule } from 'src/namespace-members/namespace-members.
     APIKeyModule,
     UserRoleModule,
     NamespacesModule,
-    NamespaceMembersModule,
     ResourcesModule,
     TasksModule,
     WizardModule,
