+++ src/namespaces/entities/namespace-member.entity.ts
@@ -1,5 +1,5 @@
 import { Base } from 'src/common/base.entity';
-import { Namespace } from 'src/namespaces/namespaces.entity';
+import { Namespace } from 'src/namespaces/entities/namespace.entity';
 import { Resource } from 'src/resources/resources.entity';
 import { User } from 'src/user/user.entity';
 import {
