+++ src/namespaces/entities/namespace.entity.ts
