+++ src/resources/resources.module.ts
@@ -4,7 +4,6 @@ import { Resource } from 'src/resources/resources.entity';
 import { ResourcesService } from 'src/resources/resources.service';
 import { ResourcesController } from 'src/resources/resources.controller';
 import { Task } from 'src/tasks/tasks.entity';
-import { NamespaceMembersModule } from 'src/namespace-members/namespace-members.module';
 import { NamespacesModule } from 'src/namespaces/namespaces.module';
 import { MinioService } from 'src/resources/minio/minio.service';
 import { InternalResourcesController } from 'src/resources/internal.resource.controller';
@@ -16,7 +15,6 @@ import { InternalResourcesController } from 'src/resources/internal.resource.con
   imports: [
     TypeOrmModule.forFeature([Resource]),
     TypeOrmModule.forFeature([Task]),
-    NamespaceMembersModule,
     NamespacesModule,
   ],
 })
